WIP Adding components in editor

+++ src/console/commands/LogEntitiesCommand.cpp
@@ -127,7 +127,7 @@ Result LogEntitiesCommand::Execute_Impl(const CommandLineArguments& args)
 
                         json::JSONObject componentJson;
                         componentJson["type"] = *componentInterface->GetTypeInfo().name;
-                        componentJson["id"] = componentId;
+                        componentJson["id"] = uint32(componentId);
 
                         if (componentTypeId == TypeId::ForType<UIComponent>())
                         {
